New class for group related contact actions
